refactor: Refactor demo app
